[Fix] Fix format string type mismatch. Found with MSVC static analyzer.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@17420 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/MIDIMappingDialog.cpp
@@ -116,7 +116,7 @@ BOOL CMIDIMappingDialog::OnInitDialog()
 	
 	// Add controller names
 	CString s;
-	for(size_t i = MIDIEvents::MIDICC_start; i <= MIDIEvents::MIDICC_end; i++)
+	for(uint8 i = MIDIEvents::MIDICC_start; i <= MIDIEvents::MIDICC_end; i++)
 	{
 		s.Format(_T("%3u "), i);
 		s += mpt::ToCString(mpt::Charset::UTF8, MIDIEvents::MidiCCNames[i]);
